requirements.txt
anyio==4.8.0
bitstruct==8.11.1
certifi==2024.12.14
charset-normalizer==3.4.1
dateutils==0.6.12
et_xmlfile==2.0.0
fit-tool==0.9.13
gpxpy==1.6.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jdcal==1.4.1
Jinja2==3.1.5
Kompy==0.0.9
MarkupSafe==3.0.2
openpyxl==2.5.12
pdoc==15.0.1
pillow==11.1.0
pocketbase==0.14.0
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
six==1.17.0
sniffio==1.3.1
urllib3==2.3.0